backend/requirements.txt
openai 
pdf2image 
boto3
requests
Pillow
